push

+++ app/views.py
@@ -276,41 +276,42 @@ def account_delete(request):
 
 
 def account_activate(request, uidb64, token):
-    try:
+    if force_str(urlsafe_base64_decode(uidb64)) and User.objects.get(pk=uid):
         uid = force_str(urlsafe_base64_decode(uidb64))
-        user = get_object_or_404(User, pk=uid)
-    except:
-        return render(request, "error-pages/404-page.html")
+        user = User.objects.get(pk=uid)
+   
 
-    if user is not None and account_activation_token.check_token(user, token):
-        user.is_active = True
-        user.save()
-        if not Personal_Details.objects.filter(user=user):
-            Personal_Details.objects.create(
-                user = user,
-                full_name = user.first_name + ' ' + user.last_name,
-                email = user.email,
-                resume_name = 'Resume 1',
-            )
-        login(request, user)
-        messages.add_message(request, messages.INFO, 'Account Successfully Activated.')
-        return redirect("/")
-    else:
-        if user:
-            user.delete()
-            messages.add_message(request, messages.INFO, 
-            """
-            Verification Authentication Timeout
-            Kindly Register Again
-            """
-            )
+        if user is not None and account_activation_token.check_token(user, token):
+            user.is_active = True
+            user.save()
+            if not Personal_Details.objects.filter(user=user):
+                Personal_Details.objects.create(
+                    user = user,
+                    full_name = user.first_name + ' ' + user.last_name,
+                    email = user.email,
+                    resume_name = 'Resume 1',
+                )
+            login(request, user)
+            messages.add_message(request, messages.INFO, 'Account Successfully Activated.')
+            return redirect("/")
         else:
-            messages.add_message(request, messages.INFO, 
-            """
-            User Does not Exist
-            """
-            )
-        return redirect("register")
+            if user:
+                user.delete()
+                messages.add_message(request, messages.INFO, 
+                """
+                Verification Authentication Timeout
+                Kindly Register Again
+                """
+                )
+            else:
+                messages.add_message(request, messages.INFO, 
+                """
+                User Does not Exist
+                """
+                )
+            return redirect("register")
+    else:
+        return render(request, "error-pages/404-page.html")
 
 
 def home(request):
